Weather Data and refactored how the routers were being exposed.
