+++ backend/src/dao.py
@@ -255,62 +255,6 @@ class Node(DAO):
                 LOGGER.error("Error executing SQL query:", e)
                 raise default_HTTP_exception(e, "insert node query")
 
-
-@dataclass
-class WeatherData(DAO):
-    """Weather Data DAO"""
-
-    wdid: int
-    tid: int
-    wdtemperature: float
-    wdhumidity: float
-    wdpressure: float
-    wddid_rain: bool
-
-    table = sql.Identifier("weatherdata")
-    id_column = sql.Identifier("wdid")
-    owner_table = sql.SQL(
-        "weatherdata NATURAL INNER JOIN timestampindex NATURAL INNER JOIN node"
-    )
-
-    @classmethod
-    async def insert(
-        cls,
-        db: AsyncConnection,
-        tid: int,
-        wdtemperature: float,
-        wdhumidity: float,
-        wdpressure: float,
-        wddid_rain: bool,
-    ):
-        async with db.cursor(row_factory=scalar_row) as curs:
-            try:
-                await curs.execute(
-                    sql.SQL(
-                        """
-INSERT INTO {table} (tid, wdtemperature, wdhumidity, wdpressure, wddid_rain)
-VALUES (%(tid)s, %(wdtemperature)s, %(wdhumidity)s, %(wdpressure)s, %(wddid_rain)s)
-RETURNING {id_column}
-"""
-                    ).format(table=cls.table, id_column=cls.id_column),
-                    {
-                        "tid": tid,
-                        "wdtemperature": wdtemperature,
-                        "wdhumidity": wdhumidity,
-                        "wdpressure": wdpressure,
-                        "wddid_rain": wddid_rain,
-                    },
-                )
-
-                db_response = await curs.fetchone()
-
-                if db_response is not None:
-                    return db_response[0]
-
-            except PGError as e:
-                LOGGER.error("Error executing SQL query:", e)
-                raise default_HTTP_exception(e, "insert weather data query")
-
 @dataclass
 class RecentData:
     """Recent reports dashboard operation response object"""
