+++ backend/src/weather/router.py
@@ -1,0 +1,2 @@
+# Re-expose router object
+from weather.service import router
