+++ backend/src/weather/service.py
@@ -1,4 +1,7 @@
-import timestamp.repository as repository
+import audio.repository as repository
+
+from weather.repository import WeatherData
+
 from dbutil import DBTransactionDependency
 from routers.security import LightWeightUser, get_current_user
 from typing import Annotated
@@ -6,18 +9,18 @@ from fastapi import Depends, Form, APIRouter
 
 router = APIRouter()
 
-@router.get("/api/timestamp/all")
-async def timestamp_all(
+@router.get("/api/weather/all")
+async def weather_all(
     current_user: Annotated[LightWeightUser, Depends(get_current_user)],
     transaction: DBTransactionDependency,
 ):
-    return await repository.TimestampIndex.get_all(current_user.auid, transaction.connection)
+    return await WeatherData.get_all(current_user.auid, transaction.connection)
 
 
-@router.get("/api/timestamp")
-async def timestamp_get(
-    tid: Annotated[int, Form()],
+@router.get("/api/weather")
+async def weather_get(
+    wdid: Annotated[int, Form()],
     current_user: Annotated[LightWeightUser, Depends(get_current_user)],
     transaction: DBTransactionDependency,
 ):
-    return await repository.TimestampIndex.get(current_user.auid, tid, transaction.connection)
+    return await WeatherData.get(current_user.auid, wdid, transaction.connection)
